requirements.txt
streamlit==1.38.0
pillow==10.4.0
numpy==2.0.0
pandas==2.2.2
scikit-learn==1.5.2
matplotlib==3.8.4

transformers==4.44.2
tokenizers==0.19.1
huggingface-hub==0.24.6
accelerate==0.34.2

torch==2.3.1+cpu
torchvision==0.18.1+cpu
--extra-index-url https://download.pytorch.org/whl/cpu

shap==0.45.1
opencv-python==4.10.0.84
tqdm==4.66.5
